actualizacion archivos grandes base64

+++ api/controllers/investigation.ts
@@ -122,13 +122,22 @@ export const putArchivo64 = async (req: Request, res: Response) => {
         investigation[0].url_archivo !== null &&
         investigation[0].url_archivo !== undefined
       ) {
-        //  console.log("si se encuentra un archivo", investigation[0].url_archivo);
-        let directory = path.join(
-          __dirname,
-          "../documents/" + investigation[0].url_archivo + ".pdf"
-        );
+        if (
+          fs.existsSync(
+            path.join(
+              __dirname,
+              "../documents/" + investigation[0].url_archivo + ".pdf"
+            )
+          )
+        ) {
+          //  console.log("si se encuentra un archivo", investigation[0].url_archivo);
+          let directory = path.join(
+            __dirname,
+            "../documents/" + investigation[0].url_archivo + ".pdf"
+          );
 
-        fs.unlinkSync(directory);
+          fs.unlinkSync(directory);
+        }
       }
       /* let encodedPdf = base64.base64Encode(
         __dirname + "/1636481213106-document-prueba.pdf"
